refactor(k8s): move retry machinery to a separate file
